requirements.txt
pandas==1.5.3
numpy==1.23.5
sqlalchemy==1.4.44
psycopg2-binary==2.9.5
requests==2.28.1